+++ packages/framework/src/Model/Payment/PaymentPriceProvider.php
@@ -2,19 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Shopsys\FrameworkBundle\Model\Cart;
+namespace Shopsys\FrameworkBundle\Model\Payment;
 
 use Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig;
+use Shopsys\FrameworkBundle\Model\Cart\Cart;
 use Shopsys\FrameworkBundle\Model\Order\Item\OrderItemTypeEnum;
 use Shopsys\FrameworkBundle\Model\Order\OrderDataFactory;
 use Shopsys\FrameworkBundle\Model\Order\Processing\OrderInputFactory;
 use Shopsys\FrameworkBundle\Model\Order\Processing\OrderProcessor;
-use Shopsys\FrameworkBundle\Model\Order\Processing\OrderProcessorMiddleware\PersonalPickupPointMiddleware;
-use Shopsys\FrameworkBundle\Model\Payment\Payment;
 use Shopsys\FrameworkBundle\Model\Pricing\Price;
-use Shopsys\FrameworkBundle\Model\Transport\Transport;
 
-class CartPriceProvider
+class PaymentPriceProvider
 {
     /**
      * @param \Shopsys\FrameworkBundle\Model\Order\Processing\OrderInputFactory $orderInputFactory
@@ -47,30 +45,5 @@ class CartPriceProvider
         );
 
         return $orderData->totalPricesByItemType[OrderItemTypeEnum::TYPE_PAYMENT];
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Cart\Cart $cart
-     * @param \Shopsys\FrameworkBundle\Model\Transport\Transport $transport
-     * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
-     * @return \Shopsys\FrameworkBundle\Model\Pricing\Price
-     */
-    public function getTransportPrice(Cart $cart, Transport $transport, DomainConfig $domainConfig): Price
-    {
-        $orderInput = $this->orderInputFactory->createFromCart($cart, $domainConfig);
-        $orderInput->setTransport($transport);
-
-        if (!$transport->isPacketery()) {
-            $orderInput->cleanAdditionalData(PersonalPickupPointMiddleware::ADDITIONAL_DATA_PICKUP_PLACE_IDENTIFIER);
-        }
-
-        $orderData = $this->orderDataFactory->create();
-
-        $orderData = $this->orderProcessor->process(
-            $orderInput,
-            $orderData,
-        );
-
-        return $orderData->totalPricesByItemType[OrderItemTypeEnum::TYPE_TRANSPORT];
     }
 }
